项目人员

+++ src/main/java/com/wondertek/cpm/web/rest/ProjectUserResource.java
@@ -1,12 +1,12 @@
 package com.wondertek.cpm.web.rest;
 
-import com.codahale.metrics.annotation.Timed;
-import com.wondertek.cpm.domain.ProjectUser;
-import com.wondertek.cpm.service.ProjectUserService;
-import com.wondertek.cpm.web.rest.util.HeaderUtil;
-import com.wondertek.cpm.web.rest.util.PaginationUtil;
+import java.net.URISyntaxException;
+import java.time.ZonedDateTime;
+import java.util.List;
+import java.util.Optional;
 
-import io.swagger.annotations.ApiParam;
+import javax.inject.Inject;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -14,17 +14,25 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.inject.Inject;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
+import com.codahale.metrics.annotation.Timed;
+import com.wondertek.cpm.config.StringUtil;
+import com.wondertek.cpm.domain.ProjectUser;
+import com.wondertek.cpm.domain.vo.ProjectUserVo;
+import com.wondertek.cpm.security.SecurityUtils;
+import com.wondertek.cpm.service.ProjectUserService;
+import com.wondertek.cpm.web.rest.util.HeaderUtil;
+import com.wondertek.cpm.web.rest.util.PaginationUtil;
 
-import static org.elasticsearch.index.query.QueryBuilders.*;
+import io.swagger.annotations.ApiParam;
 
 /**
  * REST controller for managing ProjectUser.
@@ -39,26 +47,6 @@ public class ProjectUserResource {
     private ProjectUserService projectUserService;
 
     /**
-     * POST  /project-users : Create a new projectUser.
-     *
-     * @param projectUser the projectUser to create
-     * @return the ResponseEntity with status 201 (Created) and with body the new projectUser, or with status 400 (Bad Request) if the projectUser has already an ID
-     * @throws URISyntaxException if the Location URI syntax is incorrect
-     */
-    @PostMapping("/project-users")
-    @Timed
-    public ResponseEntity<ProjectUser> createProjectUser(@RequestBody ProjectUser projectUser) throws URISyntaxException {
-        log.debug("REST request to save ProjectUser : {}", projectUser);
-        if (projectUser.getId() != null) {
-            return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("projectUser", "idexists", "A new projectUser cannot already have an ID")).body(null);
-        }
-        ProjectUser result = projectUserService.save(projectUser);
-        return ResponseEntity.created(new URI("/api/project-users/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert("projectUser", result.getId().toString()))
-            .body(result);
-    }
-
-    /**
      * PUT  /project-users : Updates an existing projectUser.
      *
      * @param projectUser the projectUser to update
@@ -69,15 +57,53 @@ public class ProjectUserResource {
      */
     @PutMapping("/project-users")
     @Timed
-    public ResponseEntity<ProjectUser> updateProjectUser(@RequestBody ProjectUser projectUser) throws URISyntaxException {
+    public ResponseEntity<Boolean> updateProjectUser(@RequestBody ProjectUser projectUser) throws URISyntaxException {
         log.debug("REST request to update ProjectUser : {}", projectUser);
-        if (projectUser.getId() == null) {
-            return createProjectUser(projectUser);
+        Boolean isNew = projectUser.getId() == null;
+        if(projectUser.getProjectId() == null || projectUser.getUserId() == null || StringUtil.isNullStr(projectUser.getUserName())
+        		 || StringUtil.isNullStr(projectUser.getUserRole()) || projectUser.getJoinDay() == null){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectUser.save.paramNone", "")).body(null);
         }
+        if(projectUser.getLeaveDay() != null && projectUser.getLeaveDay().longValue() < projectUser.getJoinDay()){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectUser.save.dayError", "")).body(null);
+        }
+        //查看用户是否被添加
+        boolean isExist = projectUserService.checkUserExist(projectUser);
+        if(isExist){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectUser.save.userIdError", "")).body(null);
+        }
+        String updator = SecurityUtils.getCurrentUserLogin();
+        ZonedDateTime updateTime = ZonedDateTime.now();
+        if(isNew){
+        	projectUser.setCreateTime(updateTime);
+        	projectUser.setCreator(updator);
+        }else{
+        	ProjectUserVo projectUserVo = projectUserService.getProjectUser(projectUser.getId());
+        	if(projectUserVo == null){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectUser.save.noPerm", "")).body(null);
+        	}
+        	ProjectUser old = projectUserService.findOne(projectUser.getId());
+        	if(old == null){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectUser.save.idNone", "")).body(null);
+        	}else if(old.getProjectId() != projectUser.getProjectId().longValue()){
+        		return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectUser.save.projectIdError", "")).body(null);
+        	}
+        	projectUser.setCreateTime(old.getCreateTime());
+        	projectUser.setCreator(old.getCreator());
+        }
+        projectUser.setUpdateTime(updateTime);
+        projectUser.setUpdator(updator);
+        
         ProjectUser result = projectUserService.save(projectUser);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert("projectUser", projectUser.getId().toString()))
-            .body(result);
+        if(isNew){
+        	return ResponseEntity.ok()
+                    .headers(HeaderUtil.createEntityCreationAlert("projectUser", result.getId().toString()))
+                    .body(isNew);
+        }else{
+        	return ResponseEntity.ok()
+        			.headers(HeaderUtil.createEntityUpdateAlert("projectUser", result.getId().toString()))
+        			.body(isNew);
+        }
     }
 
     /**
@@ -89,10 +115,19 @@ public class ProjectUserResource {
      */
     @GetMapping("/project-users")
     @Timed
-    public ResponseEntity<List<ProjectUser>> getAllProjectUsers(@ApiParam Pageable pageable)
+    public ResponseEntity<List<ProjectUserVo>> getAllProjectUsers(
+    		@RequestParam(value = "projectId") Long projectId, 
+    		@RequestParam(value = "userId") Long userId, 
+    		@ApiParam Pageable pageable)
         throws URISyntaxException {
+    	
         log.debug("REST request to get a page of ProjectUsers");
-        Page<ProjectUser> page = projectUserService.findAll(pageable);
+        ProjectUser projectUser = new ProjectUser();
+        projectUser.setProjectId(projectId);
+        projectUser.setUserId(userId);
+        
+        Page<ProjectUserVo> page = projectUserService.getUserPage(projectUser,pageable);
+        
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/project-users");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
@@ -105,10 +140,10 @@ public class ProjectUserResource {
      */
     @GetMapping("/project-users/{id}")
     @Timed
-    public ResponseEntity<ProjectUser> getProjectUser(@PathVariable Long id) {
+    public ResponseEntity<ProjectUserVo> getProjectUser(@PathVariable Long id) {
         log.debug("REST request to get ProjectUser : {}", id);
-        ProjectUser projectUser = projectUserService.findOne(id);
-        return Optional.ofNullable(projectUser)
+        ProjectUserVo projectUserVo = projectUserService.getProjectUser(id);
+        return Optional.ofNullable(projectUserVo)
             .map(result -> new ResponseEntity<>(
                 result,
                 HttpStatus.OK))
@@ -125,28 +160,13 @@ public class ProjectUserResource {
     @Timed
     public ResponseEntity<Void> deleteProjectUser(@PathVariable Long id) {
         log.debug("REST request to delete ProjectUser : {}", id);
+        ProjectUserVo projectUserVo = projectUserService.getProjectUser(id);
+        if(projectUserVo == null){
+        	return ResponseEntity.badRequest().headers(HeaderUtil.createError("cpmApp.projectUser.save.noPerm", "")).body(null);
+        }
         projectUserService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("projectUser", id.toString())).build();
     }
-
-    /**
-     * SEARCH  /_search/project-users?query=:query : search for the projectUser corresponding
-     * to the query.
-     *
-     * @param query the query of the projectUser search 
-     * @param pageable the pagination information
-     * @return the result of the search
-     * @throws URISyntaxException if there is an error to generate the pagination HTTP headers
-     */
-    @GetMapping("/_search/project-users")
-    @Timed
-    public ResponseEntity<List<ProjectUser>> searchProjectUsers(@RequestParam String query, @ApiParam Pageable pageable)
-        throws URISyntaxException {
-        log.debug("REST request to search for a page of ProjectUsers for query {}", query);
-        Page<ProjectUser> page = projectUserService.search(query, pageable);
-        HttpHeaders headers = PaginationUtil.generateSearchPaginationHttpHeaders(query, page, "/api/_search/project-users");
-        return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
-    }
-
-
+    
+    
 }
